Add files via upload
Created one module named as Arithmetic which contains 4 functions as Add() for addition, Sub()
for subtraction, Mult() for multiplication and Div() for division. All functions accepts two
parameters as number and perform the operation.A program which call all the
functions from Arithmetic module by accepting the parameters from user.